% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFextraVL.R
\name{BFchainsModelMatrix}
\alias{BFchainsModelMatrix}
\title{Model Matrix for a posterior sample chain from a BF}
\usage{
BFchainsModelMatrix(factorLevels)
}
\arguments{
\item{factorLevels}{Numerical vector which stores the number of levels for each factor in the design. The order of the factors should follow that of the elements composing the name of the highest level interaction effect in the design. For example if the highest level interaction is Etendue:levD and the Etendue and levD factors have 3 and 4 levels respectively, then the factorLevels argument should be c(3,4).}
}
\value{
The desired model matrix.
}
\description{
Returns the model matrix of the posterior sample chain of a BFmcmc object. Can be used to compute the predicted means.
}
\examples{
BFchainsModelMatrix(c(3,4))
}
