% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFextraVL.R
\name{constrBF}
\alias{constrBF}
\title{Constrained BF}
\usage{
constrBF(postDT, const, newColumnContrasts = NULL, returnData = T)
}
\arguments{
\item{postDT}{Posterior samples. Object of BFmcmc class.}

\item{const}{A character vector describing the constraints. For example, if you want to test the hypothesis according to which level A > level B and level B > level C, then const = "A>B>C". For more complicated constraints, use the semi-colon. For example if you want to test the hypothesis that level A > level B and that level A > level C, then const = "A>B;A>C". . If you want to test the two former hypotheses at once, const = c("A>B>C","A>B;A>C").}

\item{newColumnContrasts}{If supplied, this list of matrices is used to create new column from linear contrasts of the columns of the posterior samples. Each element of this list must be a matrix. The names of the columns of these matrices must correspond to column names in postDT. Rows must also be named. The row names of these matrices must be distinct from each other and from the column names of postDT.}

\item{returnData}{Boolean. Indicates whether the data table of the posterior samples (possibly with the new columns) should be included in the output or not.}
}
\value{
A list of lists with the following items:
\itemize{
\item prior Prior probability
\item post Posterior probability
\item BF_c_e Bayes factor comparing the constrained model vs the unconstrained model. 
}
}
\description{
Returns a Bayes Factor comparing a constrained model with an unconstrained one. Useful for paired comparisons.
}
\section{References}{

Morey, R. D., & Wagenmakers, E. J. (2014). Simple relation between Bayesian order-restricted and point-null hypothesis tests. Statistics and Probability Letters, 92, 121–124. http://doi.org/10.1016/j.spl.2014.05.010
Klugkist, I., Laudy, O., & Hoijtink, H. (2005). Inequality constrained analysis of variance: A Bayesian approach. Psychological Methods, 10(4), 477–493. http://doi.org/10.1037/1082-989X.10.4.477
}

\examples{
data("FPpostDT")
out<-constrBF(FPpostDT,c("Etendue-150>Etendue-900;Etendue-300>Etendue-900","levD-1<levD-2<levD-3<levD-4"))
# A more complex example.

ooo<-data.table(as.data.frame(FPpostDT))
names1<-c(paste0("levD-",1:4),grep(x=names(ooo),pattern = "^Etendue:levD",value = T,fixed = F))
mat1<-BFchainsModelMatrix(c(3,4))[,-c(1:4)]
rownames(mat1)<-paste0("c_",rep(paste0("l",1:4),times=3),"_",rep(paste0("e",c(150,300,900)),each=4))
colnames(mat1)<-names1

names2<-grep(x=names(ooo),pattern = "^Etendue:levD",value = T,fixed = F)
mat2<-kronecker(diag(1,nrow = 3),matrix(c(1,-1, 0, 0,
                                          0, 1,-1, 0,
                                          0, 0, 1, -1),nrow = 3,byrow = T))
colnames(mat2)<-names2
temp<-data.table(expand.grid(c("1-2","2-3","3-4"),c("e150","e300","e900")))
temp[,tempi:=paste0("c",Var1,"_",Var2)]
rownames(mat2)<-temp$tempi
matlist<-list(mat1,mat2)

out2<-BFextraVL::constrBF(postDT = FPpostDT,const = c("c1-2_e150<c1-2_e300<c1-2_e900;c2-3_e150<c2-3_e300<c2-3_e900","c_l1_e150<c_l2_e150<c_l3_e150<c_l4_e150;c_l1_e300<c_l2_e300<c_l3_e300<c_l4_e300"),newColumnContrasts = matlist)
}
